+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Please mark all change in change log and use the ticket from JIRA.
 
 ## Feature
 - \#12 - Pure CPU version for Milvus
+- \#77 - Support table partition
 - \#226 - Experimental shards middleware for Milvus
 
 ## Improvement